Removed unneeded buffer from generator CLI
The generator CLI code used to call talktex_header / _footer from the C
API, which was in fact not intentional. It should have been calling the
functions from the generation namespace, which work with std::strings
instead of char pointers. The old code therefore also probably had a
memory leak.

When the C API changed to work with pre-allocated buffers, the code to
call talktex_header / _footer in the CLI code also had to change, and
this change made the mixup obvious. The change involved allocating a
huge buffer, which is not at all necessary when using the generation::
functions.

This commit switches these talktex_header / _footer calls to the
generation:: variants, removing the need for a huge pre-allocated
buffer.