Removed stdout/stderr output from C API functions
It's kinda nasty to output from API functions, since it's really hard to
suppress the output from e.g. Python.
If output is required, it should be done by the user of the API. This
may possibly require returning more detailed error codes in the C API
instead of just true/false.