Added images dir and root .gitignore to release
The images are actually kinda nice to include.
As for the gitignore: the script was already including gitignores in
sudirectories anyway. We could try to filter them out, but the
gitignores are in a way part of our source code, so we might as well
just include all of them.